requirements.txt
absl-py==1.4.0
advertools==0.13.2
altair==5.1.1
appdirs==1.4.4
arrow==1.3.0
asttokens==2.4.0
astunparse==1.6.3
attrs==22.2.0
Automat==22.10.0
backcall==0.2.0
beautifulsoup4==4.12.2
blinker==1.6.2

brewer2mpl==1.4.1
bs4==0.0.1
cachetools==5.3.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.7
cloudpickle==2.2.1
colorama==0.4.6
comm==0.1.4
compress-pickle==2.1.0
constantly==15.1.0
contourpy==1.0.6
cryptography==40.0.2
cssselect==1.2.0
cycler==0.11.0
dacite==1.8.1
debugpy==1.8.0
decorator==5.1.1
entrypoints==0.4
exceptiongroup==1.1.3
executing==1.2.0
Faker==19.10.0
fastjsonschema==2.16.2
favicon==0.7.0
filelock==3.12.0
flatbuffers==23.3.3
fonttools==4.38.0
gast==0.4.0
ggplot==0.11.5
gitdb==4.0.10
GitPython==3.1.37
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.0
h5py==3.8.0
htbuilder==0.6.2
htmlmin==0.1.12
hydralit==1.0.14
hydralit-components==1.0.10
hyperlink==21.0.0
idna==3.4
ImageHash==4.3.1
importlib-metadata==6.8.0
incremental==22.10.0
ipykernel==6.25.2

itemadapter==0.8.0
itemloaders==1.1.0
jax==0.4.8
jedi==0.19.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonpatch==1.32
jsonpointer==2.3
jsonschema==4.17.3
jupyter_client==8.3.1
jupyter_core==5.3.1
keras==2.12.0
kiwisolver==1.4.4
libclang==16.0.0
llvmlite==0.41.0
lxml==4.9.2
Markdown==3.4.3
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.2
matplotlib==3.6.2
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.1.0
more-itertools==10.1.0
mpmath==1.3.0
multimethod==1.10
multitasking==0.0.11
nbformat==5.7.1
nest-asyncio==1.5.8
networkx==3.1
neuralnet==0.2.0
numba==0.58.0
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.3
pandas-profiling==3.6.6
parsel==1.8.1
parso==0.8.3
patsy==0.5.3
phik==0.12.3
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.10.0
plotly==5.11.0
prompt-toolkit==3.0.39
Protego==0.2.1
protobuf==4.23.0
psutil==5.9.5
pure-eval==0.2.2
pyarrow==12.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.13
pydeck==0.8.1b0
PyDispatcher==2.0.7
Pygments==2.16.1
pymdown-extensions==10.3
pyOpenSSL==23.1.1
pyparsing==3.0.9
pyproject-toml==0.0.10
pyrsistent==0.19.3
python-dateutil==2.8.2
python-time==0.3.0
pytz==2023.3.post1
PyWavelets==1.4.1

PyYAML==6.0.1
pyzmq==25.1.1
queuelib==1.6.2
requests==2.30.0
requests-file==1.5.1
requests-oauthlib==1.3.1
rich==13.5.3
rsa==4.9
scikit-learn==1.3.0
scipy==1.9.3
Scrapy==2.9.0
seaborn==0.12.2
service-identity==21.1.0
session-info==1.0.0
shap==0.42.1
six==1.16.0
slicer==0.0.7
smmap==5.0.1
soupsieve==2.5
st-annotated-text==4.0.1
st-mui-table==0.0.4
stack-data==0.6.2
statsmodels==0.14.0
stdlib-list==0.9.0
streamlit==1.27.2
streamlit-autorefresh==1.0.1
streamlit-camera-input-live==0.2.0
streamlit-card==0.0.61
streamlit-elements==0.1.0
streamlit-embedcode==0.1.2
streamlit-extras==0.3.4
streamlit-faker==0.0.2
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.0
streamlit-nivo==0.1.8
streamlit-option-menu==0.3.2
streamlit-pandas-profiling==0.1.3
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==1.0.2
sympy==1.12
tabulate==0.9.0
tangled-up-in-unicode==0.2.0
tenacity==8.1.0
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0

termcolor==2.3.0
Theano==1.0.5
threadpoolctl==3.1.0
tldextract==3.4.4
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tornado==6.3.3
tqdm==4.65.0
traitlets==5.10.0
treeinterpreter==0.2.3
Twisted==22.10.0

twython==3.9.1
typeguard==2.13.3
types-python-dateutil==2.8.19.14
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.2
validators==0.22.0
visdom==0.2.4
visions==0.7.5
w3lib==2.1.1
watchdog==3.0.0
wcwidth==0.2.6
websocket-client==1.5.1
wordcloud==1.9.2
wrapt==1.14.1
xgboost==1.7.3
xyzservices==2023.10.0
ydata-profiling==4.5.1
yfinance==0.1.87
zipp==3.17.0
zope.interface==6.0
